+++ Source/SwiftLintFramework/Reporters/JSONReporter.swift
@@ -11,6 +11,7 @@ import SourceKittenFramework
 
 public struct JSONReporter: Reporter {
     public static let identifier = "json"
+    public static let isRealtime = false
 
     public var description: String {
         return "Reports violations as a JSON array."