+++ Source/SwiftLintFramework/Configuration.swift
@@ -160,6 +160,7 @@ public struct Configuration {
         rules.append(NestingRule())
         rules.append(ControlStatementRule())
         rules.append(OpeningBraceRule())
+        rules.append(CommaRule())
 
         return rules
     }