+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 
 ##### Enhancements
 
+* Added `Comma Rule` to ensure there is a single space after a comma.
+  [Alex Culeva](https://github.com/S2dentik)
+
 * Add rule identifier to all linter reports.  
   [zippy1978](https://github.com/zippy1978)